Changede election to double and fixed locations string builder

+++ SMAPExam/app/src/main/java/com/grp16/itsmap/smapexam/model/google/elevation/AltResult.java
@@ -9,7 +9,7 @@ public class AltResult {
 
     @SerializedName("elevation")
     @Expose
-    public Integer elevation;
+    public double elevation;
     @SerializedName("location")
     @Expose
     public Location location;
